+++ src/whatsapp/types/wa.types.ts
@@ -22,6 +22,7 @@ export enum Events {
   GROUPS_UPSERT = 'groups.upsert',
   GROUPS_UPDATE = 'groups.update',
   GROUP_PARTICIPANTS_UPDATE = 'group-participants.update',
+  CALL = 'call',
 }
 
 export declare namespace wa {
@@ -61,6 +62,9 @@ export declare namespace wa {
     reject_call?: boolean;
     msg_call?: string;
     groups_ignore?: boolean;
+    always_online?: boolean;
+    read_messages?: boolean;
+    read_status?: boolean;
   };
 
   export type StateConnection = {