+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.4.3 (2023-07-25 10:51)
+
+### Fixed
+
+* Adjusts in settings with options always_online, read_messages and read_status
+* Fixed send webhook for event CALL
+* Create instance with settings
+
 # 1.4.2 (2023-07-24 20:52)
 
 ### Fixed